fixed buffer overflow
