+++ include/texted/insert.h
@@ -1,6 +1,11 @@
 #ifndef TEXTED_INSERT_H
 #define TEXTED_INSERT_H
 
+#include <string.h>
+
+// Empty a buffer with zeroes
+#define empty(_buffer, _size)   memset((_buffer), 0, (_size))
+
 // Insert mode
 char* insert();
 
@@ -11,8 +16,6 @@ int app_save(char* Filename, char* Buffer);
 
 // Salva il buffer nel file risrivendolo
 int save(char* Filename, char* Buffer);
-
-void empty(char* arr, int size);
 
 int streq(char* str1, char* str2, int size);
